Refactor adapters and patches
