+++ mlflow_cratedb/__init__.py
@@ -1,6 +1,6 @@
 from mlflow.utils import logging_utils
 
-from mlflow_cratedb.monkey import patch_all
+from mlflow_cratedb.boot import patch_all
 
 # Enable logging, and activate monkeypatch.
 logging_utils.enable_logging()
